Add hoop movement
Translate key presses on the client to messages to the server to move
the hoop, and send messages back to the client with the new hoop
position.

The hoop moves but things quickly crash because the client fails to read
subsequent messages. I need to fix how I read a stream of messages.