Introduce game host concept
This is a big change, but mainly it introduces the concept of a game
host per unique game that handles the connections for that game. So
a few things to make that happen:

1. A client Hello establishes which game it wants to join (just one
  hardcoded game ID for now)
2. The server processes that Hello, then either starts a host or adds it
  to an existing host.
3. Clients now get the full game state at start, then get granular game
  state updates.

The only observable effect of all this is that now if two game clients
start, the second one won't be able to control the hoop but will
see the hoop move when the first one moves it. Neat!