Proper async support to proto
I was very unhappy with my hacky async-over-sync protocol on the server,
mainly the presumption that when the asyn stream was readable it would
always have at least one message in there, so I bit the bullet and added
proper async support in proto under a feature flag so the client can
stay with no dependency on tokio.

This involved some amount of code and logic duplication between sync
and async, but overall it's much cleaner I think. Now the async path
provides a clean Stream over messages with fewer assumptions and
cleaner interface. Also this allowed me to get rid of FRAME_SIZE and
just prefix every message with its size. I did cap it to 512 though
because I got paranoid about evil clients trying to get the server to
allocate a lot. Damn evil clients.